requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Flask==2.2.5
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.6

apache-airflow==2.6.0
apache-airflow-providers-common-sql==1.4.0
apache-airflow-providers-ftp==3.3.1
apache-airflow-providers-http==4.3.0
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-postgres==5.4.0
apache-airflow-providers-sqlite==3.3.2
